requirements.txt
channels==1.1.1
coverage==4.3.1
daphne==1.1.0
dj-database-url==0.4.1
Django==1.8.1
django-allauth==0.22.0
django-autoslug==1.9.3
django-cities-light==3.3.0
django-nose==1.4.4
djangorestframework==3.5.3
gunicorn==19.6.0
MySQL-python==1.2.5
nose==1.3.7
oauthlib==2.0.0
olefile==0.44
Pillow==4.0.0
progressbar2==3.12.0
psycopg2==2.7.1
python-openid==2.2.5
python-utils==2.0.1
requests==2.12.1
requests-oauthlib==0.7.0
six==1.10.0
Unidecode==0.4.20
wheel==0.24.0
whitenoise==3.2.2
